RN-635: Broke EntityHierarchyCacher into separate classes - EntityHierarchyCacher acts as the ChangeHandler - EntityHierarchySubtreeRebuilder does the actual subtree rebuild logic
