+++ packages/database/src/changeHandlers/entityHierarchyCacher/index.js
@@ -1,0 +1,6 @@
+/**
+ * Tupaia
+ * Copyright (c) 2017 - 2022 Beyond Essential Systems Pty Ltd
+ */
+
+export { EntityHierarchyCacher } from './EntityHierarchyCacher';
