+++ packages/database/src/index.js
@@ -20,5 +20,5 @@ export { DatabaseChangeChannel } from './DatabaseChangeChannel';
 export { DatabaseModel } from './DatabaseModel';
 export { DatabaseType } from './DatabaseType';
 export * from './testUtilities';
-export { getConnectionConfig } from  './getConnectionConfig';
+export { getConnectionConfig } from './getConnectionConfig';
 export { SqlQuery } from './SqlQuery';
